+++ app/views/DefaultBrowserView/index.tsx
@@ -1,21 +1,21 @@
-import React, { memo, useCallback, useEffect, useLayoutEffect, useState } from 'react';
+import React, { useCallback, useEffect, useLayoutEffect, useState } from 'react';
 import { FlatList, Linking } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 
-import I18n from '../i18n';
-import { TSupportedThemes, useTheme } from '../theme';
-import { themes } from '../lib/constants';
-import StatusBar from '../containers/StatusBar';
-import * as List from '../containers/List';
-import { DEFAULT_BROWSER_KEY } from '../lib/methods/helpers/openLink';
-import { isIOS } from '../lib/methods/helpers';
-import SafeAreaView from '../containers/SafeAreaView';
-import UserPreferences from '../lib/methods/userPreferences';
-import { events, logEvent } from '../lib/methods/helpers/log';
+import I18n from '../../i18n';
+import { useTheme } from '../../theme';
+import StatusBar from '../../containers/StatusBar';
+import * as List from '../../containers/List';
+import { DEFAULT_BROWSER_KEY } from '../../lib/methods/helpers/openLink';
+import { isIOS } from '../../lib/methods/helpers';
+import SafeAreaView from '../../containers/SafeAreaView';
+import UserPreferences from '../../lib/methods/userPreferences';
+import { events, logEvent } from '../../lib/methods/helpers/log';
+import Item from './Item';
 
-type TValue = 'inApp' | 'systemDefault:' | 'googlechrome:' | 'firefox:' | 'brave:';
+export type TValue = 'inApp' | 'systemDefault:' | 'googlechrome:' | 'firefox:' | 'brave:';
 
-interface IBrowsersValues {
+export interface IBrowsersValues {
 	title: string;
 	value: TValue;
 }
@@ -46,33 +46,8 @@ const BROWSERS: IBrowsersValues[] = [
 	}
 ];
 
-interface IRenderItem extends IBrowsersValues {
-	browser: string;
-	theme: TSupportedThemes;
-	changeDefaultBrowser: (newBrowser: TValue) => void;
-}
-
-const RenderItem = memo(({ title, value, browser, changeDefaultBrowser, theme }: IRenderItem) => {
-	let isSelected = false;
-	if (!browser && value === 'systemDefault:') {
-		isSelected = true;
-	} else {
-		isSelected = browser === value;
-	}
-
-	return (
-		<List.Item
-			title={I18n.t(title, { defaultValue: title })}
-			onPress={() => changeDefaultBrowser(value)}
-			testID={`default-browser-view-${title}`}
-			right={() => (isSelected ? <List.Icon name='check' color={themes[theme].tintColor} /> : null)}
-			translateTitle={false}
-		/>
-	);
-});
-
 const DefaultBrowserView = () => {
-	const [browser, setBrowser] = useState<string>('');
+	const [browser, setBrowser] = useState<string | null>(null);
 	const [supported, setSupported] = useState<IBrowsersValues[]>([]);
 
 	const { theme } = useTheme();
@@ -85,6 +60,9 @@ const DefaultBrowserView = () => {
 	}, [navigation]);
 
 	useEffect(() => {
+		const getBrowser = UserPreferences.getString(DEFAULT_BROWSER_KEY);
+		setBrowser(getBrowser);
+
 		if (isIOS) {
 			BROWSERS.forEach(browser => {
 				const { value } = browser;
@@ -116,7 +94,7 @@ const DefaultBrowserView = () => {
 				keyExtractor={item => item.value}
 				contentContainerStyle={List.styles.contentContainerStyleFlatList}
 				renderItem={({ item }) => (
-					<RenderItem
+					<Item
 						browser={browser}
 						theme={theme}
 						changeDefaultBrowser={changeDefaultBrowser}
